added invoicePdf

+++ src/main/java/com/viser/StockTrade/dto/PurchasesItemDto.java
@@ -11,9 +11,9 @@ public class PurchasesItemDto {
     private String product;
     @Min(value = 1, message = "Quantity must be at least 1")
     private int quantity;
-    private int price;
+    private double price;
 
-    public PurchasesItemDto(String product, int quantity, int price) {
+    public PurchasesItemDto(String product, int quantity, double price) {
         this.product = product;
         this.quantity = quantity;
         this.price = price;
